Add FileInvoker
This invoker can be used instead of the SudoInvoker to not have to
actually call `smartctl` on the local machine. We could use this in
production to decouple the part that talks to hardware and the part that
answers random requests from the network, but it's main purpose is
testing and development.