smartctl: Ignore interpreted disk status code
Only the lower 3 bits of the return code signify an error with the
application. The higher bits all represent the disk status, which we are
not interested in.

I had a disk with a reported error in the S.M.A.R.T. log which resulted
in all smartctl invocations returning with exit coe 64.